replaced precisions

+++ include/rayslib.glsl
@@ -59,7 +59,7 @@ layout ( std430, binding = 11, set = 0 ) buffer ispaceB { uint16_t ispace[][R_BL
 #define m16i(b,i) (int(ispace[b][i])-1)
 #define m16s(a,b,i) (ispace[b][i] = uint16_t(a+1))
 #else
-layout ( std430, binding = 11, set = 0 ) buffer ispaceB { mediump uint ispace[][R_BLOCK_SIZE]; };
+layout ( std430, binding = 11, set = 0 ) buffer ispaceB { highp uint ispace[][R_BLOCK_SIZE]; };
 #define m16i(b,i) (int(ispace[b][i])-1)
 #define m16s(a,b,i) (ispace[b][i] = uint(a+1))
 #endif
